backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
numpy>=1.26.0
python-multipart
google-generativeai>=0.7.0
# statistics module is part of Python stdlib — no extra install needed
